# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BeginRegistrationRequest {
  /** The Bitcoin address to bind to the biometric credential */
  bitcoinAddress: string;
}

export type RegistrationOptionsRp = {
  name: string;
  id: string;
};

export type RegistrationOptionsUser = {
  id: string;
  name: string;
  displayName: string;
};

export type RegistrationOptionsPubKeyCredParamsItem = {
  type: string;
  alg: number;
};

export type RegistrationOptionsAuthenticatorSelection = { [key: string]: unknown };

export type RegistrationOptionsExcludeCredentialsItem = { [key: string]: unknown };

/**
 * WebAuthn PublicKeyCredentialCreationOptions (JSON-serializable form)
 */
export interface RegistrationOptions {
  /** Base64url-encoded challenge */
  challenge: string;
  rp: RegistrationOptionsRp;
  user: RegistrationOptionsUser;
  pubKeyCredParams: RegistrationOptionsPubKeyCredParamsItem[];
  authenticatorSelection?: RegistrationOptionsAuthenticatorSelection;
  attestation?: string;
  timeout?: number;
  excludeCredentials?: RegistrationOptionsExcludeCredentialsItem[];
}

export interface MockRegisterRequest {
  /** Synthetic credential ID (random hex UUID) */
  credentialId: string;
  /** SPKI-format P-256 public key, base64-encoded, generated by SubtleCrypto */
  publicKeySpki: string;
  /** Optional link token — when present this account is created as Node B (eval step) */
  linkToken?: string;
}

export interface MockRegisterResult {
  vaultId: string;
  credentialId: string;
  linkedCommitment?: string | null;
  message: string;
}

export interface GenerateLinkRequest {
  /** The vault ID of Account A — the originator that will generate the link token */
  bitcoinAddress: string;
}

export interface GenerateLinkResult {
  /** One-time token to share with Account B. Represents the partially applied function — Account B's fingerprint scan completes ("evals") it into a bound address.
   */
  linkToken: string;
  /** sha256(linkToken) — the anonymous shared commitment stored by both accounts */
  commitment: string;
  /** The vault ID that Account B will receive when they register with this token */
  derivedAccountId: string;
  expiresAt: string;
}

export type CompleteRegistrationRequestAttestationResponseResponse = {
  clientDataJSON: string;
  attestationObject: string;
};

/**
 * The raw JSON credential object from navigator.credentials.create()
 */
export type CompleteRegistrationRequestAttestationResponse = {
  id: string;
  rawId: string;
  type: string;
  response: CompleteRegistrationRequestAttestationResponseResponse;
};

export interface CompleteRegistrationRequest {
  bitcoinAddress: string;
  /** Base64url-encoded challenge generated client-side. If provided, the server skips the DB challenge lookup and verifies directly. */
  challenge?: string;
  /** Optional. When provided, Account B is created as a contingent pair of the account that generated this token. The vault ID is derived from the token's commitment — eval(apply(credentialA)) = addressB.
   */
  linkToken?: string;
  /** The raw JSON credential object from navigator.credentials.create() */
  attestationResponse: CompleteRegistrationRequestAttestationResponse;
}

export interface RegistrationResult {
  success: boolean;
  /** The registered credential ID bound to the Bitcoin address */
  credentialId: string;
  bitcoinAddress: string;
  registeredAt: string;
  message: string;
}

export interface BeginAuthenticationRequest {
  /** The Bitcoin address to prove ownership of */
  bitcoinAddress: string;
}

export type AuthenticationOptionsAllowCredentialsItem = { [key: string]: unknown };

/**
 * WebAuthn PublicKeyCredentialRequestOptions (JSON-serializable form)
 */
export interface AuthenticationOptions {
  challenge: string;
  allowCredentials?: AuthenticationOptionsAllowCredentialsItem[];
  userVerification?: string;
  timeout?: number;
  rpId?: string;
}

export type CompleteAuthenticationRequestAssertionResponseResponse = {
  clientDataJSON: string;
  authenticatorData: string;
  signature: string;
  userHandle?: string;
};

/**
 * The raw JSON credential object from navigator.credentials.get()
 */
export type CompleteAuthenticationRequestAssertionResponse = {
  id: string;
  rawId: string;
  type: string;
  response: CompleteAuthenticationRequestAssertionResponseResponse;
};

export interface CompleteAuthenticationRequest {
  bitcoinAddress: string;
  /** The raw JSON credential object from navigator.credentials.get() */
  assertionResponse: CompleteAuthenticationRequestAssertionResponse;
}

export interface AuthenticationResult {
  success: boolean;
  bitcoinAddress: string;
  credentialId: string;
  provenAt: string;
  message: string;
  /** Updated authenticator sign counter (replay attack detection) */
  newSignCount?: number;
}

export interface CredentialSummary {
  id: number;
  bitcoinAddress: string;
  credentialId: string;
  registeredAt: string;
  lastUsedAt?: string | null;
  signCount: number;
  /** sha256 commitment shared with the contingent pair account. Non-null means this account is one half of an anonymous linked pair.
   */
  linkedCommitment?: string | null;
}

export interface CredentialList {
  credentials: CredentialSummary[];
  total: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}