# lib/api-zod/src/generated/types/authenticationOptions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthenticationOptionsAllowCredentialsItem } from './authenticationOptionsAllowCredentialsItem';

/**
 * WebAuthn PublicKeyCredentialRequestOptions (JSON-serializable form)
 */
export interface AuthenticationOptions {
  challenge: string;
  allowCredentials?: AuthenticationOptionsAllowCredentialsItem[];
  userVerification?: string;
  timeout?: number;
  rpId?: string;
}