# lib/api-zod/src/generated/types/authenticationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthenticationResult {
  success: boolean;
  bitcoinAddress: string;
  credentialId: string;
  provenAt: Date;
  message: string;
  /** Updated authenticator sign counter (replay attack detection) */
  newSignCount?: number;
}